Added: Steps component doc.
