+++ content/docs/components/steps.md
@@ -1,0 +1,34 @@
+---
+title: "Steps"
+description: "Steps Component is a user interface element used to guide users through a multi-step process or workflow. It presents a visual representation of the various stages or steps a user needs to complete in a sequential order. Each step is typically accompanied by a title, description, and an indicator of completion status, helping users track their progress."
+---
+
+## Default Step
+
+The Default Step refers to the fundamental and default configuration of a Steps Component within a user interface. It provides a clear visual representation of each step in a multi-step process.
+::DefaultStep
+::
+
+## Step With Border
+
+The `border` variant of the Step Point adds a border around the step indicator, making it more prominent and visually distinct.
+::StepWithBorder
+::
+
+## Step With Icon
+
+The `icon` variant of the Step Point includes an icon within the step indicator, providing an additional visual cue.
+::StepWithIcon
+::
+
+## Step Point API Reference
+
+Explore the available props for the `<StepPoint/>` component.
+::StepPointApi
+::
+
+## Step Line API Reference
+
+Explore the available props for the `<StepLine/>` component.
+::StepLineApi
+::
